Make JoinRow a top-level class instead of an inner class of JoinRowMapper
